requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.8
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.41.0
django-cors-headers==3.2.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
oauthlib==3.1.0
pep8==1.7.1
Pillow==7.0.0
pycodestyle==2.5.0
pycparser==2.20
pylint==2.4.4
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
urllib3==1.25.8
whitenoise==5.0.1
wrapt
